Unify build tool plugins into single SafeDIGenerator with artifact bundle support
Replace the dual-plugin architecture (SafeDIGenerator + SafeDIPrebuiltGenerator) with a
single SafeDIGenerator plugin that shells out to SafeDITool for scanning and generation.

- Add `scan` and `generate` subcommands to SafeDITool (`generate` is default for
  backward compatibility)
- Plugin calls `SafeDITool scan` via Process during createBuildCommands to build
  the manifest, then returns a `.buildCommand` for `SafeDITool generate`
- Add `usePrebuiltBinary` flag in Package.swift to switch between artifact bundle
  (prebuilt) and source-built SafeDITool
- Add artifact bundle binary target (placeholder checksum, updated by publish workflow)
- Move output file naming and relative path utilities from SafeDIScannerCore to SafeDICore
- Add `additionalInputFiles` field to SafeDIToolManifest for build input tracking
- Delete SafeDIScannerCore, SafeDIScanner, SafeDIPrebuiltGenerator, InstallSafeDITool
- Update publish workflow: workflow_dispatch with version/branch/dry-run inputs,
  Linux builds, artifact bundle assembly, auto-update Package.swift
- Update documentation, examples, and CI
