--- Plugins/SafeDIGenerator/SafeDIScanner.swift
@@ -1,1 +1,0 @@
-../../Sources/SafeDIScannerCore/SafeDIScanner.swift
