--- Plugins/SafeDIGenerator/SharedSafeDIScanner.swift
@@ -1,1 +1,0 @@
-../SharedSafeDIScanner.swift
