--- Plugins/SafeDIPrebuiltGenerator/RelativePath.swift
@@ -1,1 +1,0 @@
-../../Sources/SafeDIScannerCore/RelativePath.swift
