--- Plugins/SafeDIPrebuiltGenerator/SafeDIGenerateDependencyTree.swift
@@ -1,241 +1,0 @@
-// Distributed under the MIT License
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-import Foundation
-import PackagePlugin
-
-@main
-struct SafeDIGenerateDependencyTree: BuildToolPlugin {
-	func createBuildCommands(
-		context: PluginContext,
-		target: Target,
-	) async throws -> [Command] {
-		guard let sourceTarget = target as? SourceModuleTarget else {
-			return []
-		}
-
-		let outputDirectory = context.pluginWorkDirectoryURL.appending(path: "SafeDIOutput")
-		// Swift Package Plugins did not (as of Swift 5.9) allow for
-		// creating dependencies between plugin output at the time of writing.
-		// Since our current build system did not support depending on the
-		// output of other plugins, we must forgo searching for `.safeDI` files
-		// and instead parse the entire project at once.
-		// TODO: https://github.com/dfed/SafeDI/issues/92
-		let targetSwiftFiles = sourceTarget.sourceFiles(withSuffix: ".swift").map(\.url)
-		let dependenciesSourceFiles = sourceTarget
-			.sourceModuleRecursiveDependencies
-			.flatMap {
-				$0
-					.sourceFiles(withSuffix: ".swift")
-					.map(\.url)
-			}
-
-		let allSwiftFiles = targetSwiftFiles + dependenciesSourceFiles
-		let packageRoot = context.package.directoryURL
-		let inputSourcesFile = context.pluginWorkDirectoryURL.appending(path: "InputSwiftFiles.csv")
-		try writeInputSwiftFilesCSV(
-			allSwiftFiles,
-			relativeTo: packageRoot,
-			to: inputSourcesFile,
-		)
-
-		let additionalSwiftFiles = discoverAdditionalDirectorySwiftFiles(
-			in: targetSwiftFiles,
-			relativeTo: packageRoot,
-		)
-
-		let manifestFile = context.pluginWorkDirectoryURL.appending(path: "SafeDIManifest.json")
-		let scanResult = try runSafeDIScanner(
-			inputSourcesFile: inputSourcesFile,
-			projectRoot: packageRoot,
-			outputDirectory: outputDirectory,
-			manifestFile: manifestFile,
-			additionalSwiftFiles: additionalSwiftFiles,
-			mockScopedSwiftFiles: targetSwiftFiles,
-		)
-		guard !scanResult.outputFiles.isEmpty else {
-			return []
-		}
-
-		let arguments = [
-			inputSourcesFile.path(percentEncoded: false),
-			"--swift-manifest",
-			manifestFile.path(percentEncoded: false),
-		]
-
-		let downloadedToolLocation = context.downloadedToolLocation
-		let safeDIVersion = context.safeDIVersion
-
-		let toolLocation = if let downloadedToolLocation {
-			downloadedToolLocation
-		} else if let safeDIVersion {
-			struct NoReleaseBinaryFoundError: Error, CustomStringConvertible {
-				let safeDIVersion: String
-				let packagePath: String
-
-				var description: String {
-					"""
-					Install the release SafeDITool binary for version \(safeDIVersion):
-					\tswift package --package-path "\(packagePath)" --allow-network-connections all --allow-writing-to-package-directory safedi-release-install
-					"""
-				}
-			}
-			throw NoReleaseBinaryFoundError(safeDIVersion: safeDIVersion, packagePath: context.package.directoryURL.path(percentEncoded: false))
-		} else {
-			struct NoReleaseBinaryAvailableError: Error, CustomStringConvertible {
-				var description: String {
-					"SafeDIPrebuiltGenerator is only supported when using a versioned release."
-				}
-			}
-			throw NoReleaseBinaryAvailableError()
-		}
-
-		return [
-			.buildCommand(
-				displayName: "SafeDIGenerateDependencyTree",
-				executable: toolLocation,
-				arguments: arguments,
-				environment: [:],
-				inputFiles: allSwiftFiles + scanResult.additionalInputFiles,
-				outputFiles: scanResult.outputFiles,
-			),
-		]
-	}
-}
-
-extension Target {
-	var sourceModuleRecursiveDependencies: [SwiftSourceModuleTarget] {
-		recursiveTargetDependencies.compactMap { target in
-			// Since we only understand Swift files, we only care about SwiftSourceModuleTargets.
-			guard let swiftModule = target as? SwiftSourceModuleTarget else {
-				return nil
-			}
-
-			// We only care about first-party code. Ignore third-party dependencies.
-			guard swiftModule
-				.directoryURL
-				.pathComponents
-				// Removing the module name.
-				.dropLast()
-				// Removing 'Sources'.
-				.dropLast(ifEquals: "Sources")
-				// Removing the package name.
-				.dropLast()
-				.last != "checkouts"
-			else {
-				return nil
-			}
-			return swiftModule
-		}
-	}
-}
-
-#if canImport(XcodeProjectPlugin)
-	import XcodeProjectPlugin
-
-	extension SafeDIGenerateDependencyTree: XcodeBuildToolPlugin {
-		func createBuildCommands(
-			context: XcodeProjectPlugin.XcodePluginContext,
-			target: XcodeProjectPlugin.XcodeTarget,
-		) throws -> [PackagePlugin.Command] {
-			// As of Xcode 15.0.1, Swift Package Plugins in Xcode are unable
-			// to inspect target dependencies. As a result, this Xcode plugin
-			// only works if it is running on a single-module project, or if
-			// all `@Instantiable`-decorated types are in the target module,
-			// or if a #SafeDIConfiguration's `additionalDirectoriesToInclude`
-			// directs the plugin to search additional modules for Swift files.
-			// https://github.com/apple/swift-package-manager/issues/6003
-			let inputSwiftFiles = target
-				.inputFiles
-				.filter { $0.url.pathExtension == "swift" }
-				.map(\.url)
-			guard !inputSwiftFiles.isEmpty else {
-				// There are no Swift files in this module!
-				return []
-			}
-
-			let outputDirectory = context.pluginWorkDirectoryURL.appending(path: "SafeDIOutput")
-			let projectRoot = context.xcodeProject.directoryURL
-			let inputSourcesFile = context.pluginWorkDirectoryURL.appending(path: "InputSwiftFiles.csv")
-			try writeInputSwiftFilesCSV(
-				inputSwiftFiles,
-				relativeTo: projectRoot,
-				to: inputSourcesFile,
-			)
-
-			let additionalSwiftFiles = discoverAdditionalDirectorySwiftFiles(
-				in: inputSwiftFiles,
-				relativeTo: projectRoot,
-			)
-
-			let manifestFile = context.pluginWorkDirectoryURL.appending(path: "SafeDIManifest.json")
-			let scanResult = try runSafeDIScanner(
-				inputSourcesFile: inputSourcesFile,
-				projectRoot: projectRoot,
-				outputDirectory: outputDirectory,
-				manifestFile: manifestFile,
-				additionalSwiftFiles: additionalSwiftFiles,
-				mockScopedSwiftFiles: inputSwiftFiles,
-			)
-			guard !scanResult.outputFiles.isEmpty else {
-				return []
-			}
-
-			let arguments = [
-				inputSourcesFile.path(percentEncoded: false),
-				"--swift-manifest",
-				manifestFile.path(percentEncoded: false),
-			]
-
-			let downloadedToolLocation = context.downloadedToolLocation
-			let toolLocation = if let downloadedToolLocation {
-				downloadedToolLocation
-			} else {
-				struct NoReleaseBinaryFoundError: Error, CustomStringConvertible {
-					var description: String {
-						"To install the release SafeDITool binary for this version, run the `InstallSafeDITool` command plugin."
-					}
-				}
-				throw NoReleaseBinaryFoundError()
-			}
-
-			return [
-				.buildCommand(
-					displayName: "SafeDIGenerateDependencyTree",
-					executable: toolLocation,
-					arguments: arguments,
-					environment: [:],
-					inputFiles: inputSwiftFiles + scanResult.additionalInputFiles,
-					outputFiles: scanResult.outputFiles,
-				),
-			]
-		}
-	}
-#endif
-
-extension Array where Element: Equatable {
-	public func dropLast(ifEquals value: Element) -> [Element] {
-		if last == value {
-			dropLast()
-		} else {
-			self
-		}
-	}
-}
