--- Plugins/SafeDIPrebuiltGenerator/Shared.swift
@@ -1,1 +1,0 @@
-../Shared.swift
