--- Sources/SafeDIScanner/SafeDIScannerCommand.swift
@@ -1,49 +1,0 @@
-// Distributed under the MIT License
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-import ArgumentParser
-import Foundation
-import SafeDIScannerCore
-
-@main
-struct SafeDIScannerCommand: ParsableCommand {
-	@Option(help: "A path to a CSV file containing paths of Swift files to scan.") var inputSourcesFile: String
-
-	@Option(help: "The root directory of the project, used to compute relative paths.") var projectRoot: String
-
-	@Option(help: "The directory where generated output files will be written.") var outputDirectory: String
-
-	@Option(help: "The path where the manifest JSON file will be written.") var manifestFile: String
-
-	func run() throws {
-		let scanner = SafeDIScanner()
-		let inputSourcesFileURL = URL(fileURLWithPath: inputSourcesFile)
-		let projectRootURL = URL(fileURLWithPath: projectRoot)
-		let outputDirectoryURL = URL(fileURLWithPath: outputDirectory)
-		let manifestFileURL = URL(fileURLWithPath: manifestFile)
-		let inputFilePaths = try SafeDIScanner.inputFilePaths(from: inputSourcesFileURL)
-		let result = try scanner.scan(
-			inputFilePaths: inputFilePaths,
-			relativeTo: projectRootURL,
-			outputDirectory: outputDirectoryURL,
-		)
-		try result.writeManifest(to: manifestFileURL)
-	}
-}
