--- Sources/SafeDIScannerCore/SafeDIScanner.swift
@@ -1,851 +1,0 @@
-// Distributed under the MIT License
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-import Foundation
-
-public struct SafeDIScanner {
-	public struct Manifest: Codable, Equatable {
-		public struct InputOutputMap: Codable, Equatable {
-			// These field names must stay in sync with SafeDIToolManifest.InputOutputMap.
-			public var inputFilePath: String
-			public var outputFilePath: String
-
-			public init(inputFilePath: String, outputFilePath: String) {
-				self.inputFilePath = inputFilePath
-				self.outputFilePath = outputFilePath
-			}
-		}
-
-		public var dependencyTreeGeneration: [InputOutputMap]
-		public var mockGeneration: [InputOutputMap]
-		public var configurationFilePaths: [String]
-		public var mockConfigurationOutputFilePath: String?
-		public var additionalMocksToGenerate: [String]
-
-		public init(
-			dependencyTreeGeneration: [InputOutputMap],
-			mockGeneration: [InputOutputMap],
-			configurationFilePaths: [String] = [],
-			mockConfigurationOutputFilePath: String? = nil,
-			additionalMocksToGenerate: [String] = [],
-		) {
-			self.dependencyTreeGeneration = dependencyTreeGeneration
-			self.mockGeneration = mockGeneration
-			self.configurationFilePaths = configurationFilePaths
-			self.mockConfigurationOutputFilePath = mockConfigurationOutputFilePath
-			self.additionalMocksToGenerate = additionalMocksToGenerate
-		}
-	}
-
-	public struct Result: Equatable {
-		public init(manifest: Manifest) {
-			self.manifest = manifest
-		}
-
-		public let manifest: Manifest
-
-		public var outputFiles: [URL] {
-			var files = (manifest.dependencyTreeGeneration + manifest.mockGeneration).map {
-				URL(fileURLWithPath: $0.outputFilePath)
-			}
-			if let mockConfigurationOutputFilePath = manifest.mockConfigurationOutputFilePath {
-				files.append(URL(fileURLWithPath: mockConfigurationOutputFilePath))
-			}
-			return files
-		}
-
-		public func writeManifest(to manifestURL: URL) throws {
-			let encoder = JSONEncoder()
-			encoder.outputFormatting = [.sortedKeys]
-			try encoder.encode(manifest).write(to: manifestURL)
-		}
-	}
-
-	public init() {}
-
-	public func scan(
-		inputFilePaths: [String],
-		relativeTo baseURL: URL,
-		outputDirectory: URL,
-	) throws -> Result {
-		let directoryBaseURL = baseURL.hasDirectoryPath
-			? baseURL
-			: baseURL.appendingPathComponent("", isDirectory: true)
-		let allFiles = inputFilePaths.map { inputFilePath in
-			URL(fileURLWithPath: inputFilePath, relativeTo: directoryBaseURL).standardizedFileURL
-		}
-		return try scan(
-			swiftFiles: allFiles,
-			targetSwiftFiles: allFiles,
-			relativeTo: baseURL,
-			outputDirectory: outputDirectory,
-		)
-	}
-
-	/// - Parameters:
-	///   - swiftFiles: All swift files to scan (target + dependencies) for root detection.
-	///   - targetSwiftFiles: Only the target module's swift files, for mock generation scoping.
-	///   - baseURL: The base URL for computing relative paths.
-	///   - outputDirectory: Where to write output files.
-	public func scan(
-		swiftFiles: [URL],
-		targetSwiftFiles: [URL]? = nil,
-		relativeTo baseURL: URL,
-		outputDirectory: URL,
-	) throws -> Result {
-		let sortedSwiftFiles = swiftFiles.sorted {
-			relativePath(for: $0, relativeTo: baseURL) < relativePath(for: $1, relativeTo: baseURL)
-		}
-		let rootFiles = try sortedSwiftFiles.filter(Self.fileContainsRoot(at:))
-		let rootOutputFileNames = Self.outputFileNames(for: rootFiles, relativeTo: baseURL)
-
-		// Mock generation is scoped to target files only (to avoid duplicates in multi-module builds).
-		let filesForMockScan = (targetSwiftFiles ?? swiftFiles).sorted {
-			relativePath(for: $0, relativeTo: baseURL) < relativePath(for: $1, relativeTo: baseURL)
-		}
-
-		// Find config files in the target's files (not dependency files).
-		let configurationFilePaths = try filesForMockScan
-			.filter(Self.fileContainsConfiguration(at:))
-			.map { relativePath(for: $0, relativeTo: baseURL) }
-
-		// Only create mock output entries for files that contain @Instantiable(generateMock: true).
-		let instantiableFiles = try filesForMockScan.filter { fileURL in
-			let source = try String(contentsOf: fileURL, encoding: .utf8)
-			return Self.containsGenerateMockTrue(in: source)
-		}
-		let mockOutputFileNames = Self.mockOutputFileNames(for: instantiableFiles, relativeTo: baseURL)
-
-		// Extract additionalMocksToGenerate from the config file (if any).
-		let additionalMocksToGenerate: [String] = {
-			for fileURL in filesForMockScan {
-				guard let content = try? String(contentsOf: fileURL, encoding: .utf8) else { continue }
-				let mocks = Self.extractAdditionalMocksToGenerate(in: content)
-				guard !mocks.isEmpty else { continue }
-				return mocks
-			}
-			return []
-		}()
-
-		// Generate mock output entries for additional mock types.
-		let additionalMockOutputEntries: [Manifest.InputOutputMap] = additionalMocksToGenerate.map { typeName in
-			.init(
-				inputFilePath: typeName,
-				outputFilePath: outputDirectory
-					.appendingPathComponent("\(typeName)+SafeDIMock.swift")
-					.path,
-			)
-		}
-
-		let hasMockEntries = !instantiableFiles.isEmpty || !additionalMockOutputEntries.isEmpty
-		let mockConfigurationOutputFilePath: String? = if hasMockEntries {
-			outputDirectory
-				.appendingPathComponent("SafeDIMockConfiguration.swift")
-				.path
-		} else {
-			nil
-		}
-
-		return Result(
-			manifest: Manifest(
-				dependencyTreeGeneration: zip(rootFiles, rootOutputFileNames).map { inputURL, outputFileName in
-					.init(
-						inputFilePath: relativePath(for: inputURL, relativeTo: baseURL),
-						outputFilePath: outputDirectory
-							.appendingPathComponent(outputFileName)
-							.path,
-					)
-				},
-				mockGeneration: zip(instantiableFiles, mockOutputFileNames).map { inputURL, outputFileName in
-					.init(
-						inputFilePath: relativePath(for: inputURL, relativeTo: baseURL),
-						outputFilePath: outputDirectory
-							.appendingPathComponent(outputFileName)
-							.path,
-					)
-				} + additionalMockOutputEntries,
-				configurationFilePaths: configurationFilePaths,
-				mockConfigurationOutputFilePath: mockConfigurationOutputFilePath,
-				additionalMocksToGenerate: additionalMocksToGenerate,
-			),
-		)
-	}
-
-	public static func inputFilePaths(from csvURL: URL) throws -> [String] {
-		try String(contentsOf: csvURL, encoding: .utf8)
-			.components(separatedBy: CharacterSet(arrayLiteral: ","))
-			.filter { !$0.isEmpty }
-	}
-
-	public static func fileContainsRoot(at fileURL: URL) throws -> Bool {
-		containsRoot(in: try String(contentsOf: fileURL, encoding: .utf8))
-	}
-
-	public static func fileContainsConfiguration(at fileURL: URL) throws -> Bool {
-		containsConfiguration(in: try String(contentsOf: fileURL, encoding: .utf8))
-	}
-
-	public static func containsConfiguration(in source: String) -> Bool {
-		let sanitizedSource = sanitize(source: source)
-		let macroName = "#SafeDIConfiguration"
-		var searchStart = sanitizedSource.startIndex
-		while let range = sanitizedSource[searchStart...].range(of: macroName) {
-			let afterMacro = range.upperBound
-			if afterMacro >= sanitizedSource.endIndex || !isIdentifierContinuation(sanitizedSource[afterMacro]) {
-				return true
-			} else {
-				searchStart = afterMacro
-			}
-		}
-		return false
-	}
-
-	public static func containsInstantiable(in source: String) -> Bool {
-		let sanitizedSource = sanitize(source: source)
-		let macroName = "@Instantiable"
-		var searchStart = sanitizedSource.startIndex
-
-		while let macroRange = sanitizedSource[searchStart...].range(of: macroName) {
-			let index = macroRange.upperBound
-			if index < sanitizedSource.endIndex,
-			   isIdentifierContinuation(sanitizedSource[index])
-			{
-				searchStart = index
-				continue
-			}
-			// Found a valid @Instantiable token
-			return true
-		}
-
-		return false
-	}
-
-	public static func containsRoot(in source: String) -> Bool {
-		let sanitizedSource = sanitize(source: source)
-		let macroName = "@Instantiable"
-		var searchStart = sanitizedSource.startIndex
-
-		while let macroRange = sanitizedSource[searchStart...].range(of: macroName) {
-			var index = macroRange.upperBound
-			if index < sanitizedSource.endIndex,
-			   isIdentifierContinuation(sanitizedSource[index])
-			{
-				searchStart = index
-				continue
-			}
-
-			skipWhitespace(in: sanitizedSource, index: &index)
-			guard index < sanitizedSource.endIndex,
-			      sanitizedSource[index] == "(",
-			      let closingParenIndex = matchingParenIndex(
-			      	in: sanitizedSource,
-			      	openingParenIndex: index,
-			      )
-			else {
-				searchStart = macroRange.upperBound
-				continue
-			}
-
-			let arguments = sanitizedSource[sanitizedSource.index(after: index)..<closingParenIndex]
-			if containsRootArgument(in: arguments) {
-				return true
-			}
-
-			searchStart = sanitizedSource.index(after: closingParenIndex)
-		}
-
-		return false
-	}
-
-	private static func mockOutputFileNames(
-		for inputURLs: [URL],
-		relativeTo baseURL: URL,
-	) -> [String] {
-		outputFileNames(for: inputURLs, relativeTo: baseURL, suffix: "+SafeDIMock.swift")
-	}
-
-	/// Extracts `additionalMocksToGenerate` type names from a source file
-	/// containing `#SafeDIConfiguration`. Uses text-based scanning (no SwiftSyntax).
-	/// Returns an empty array if the file does not contain a configuration.
-	public static func extractAdditionalMocksToGenerate(in source: String) -> [String] {
-		extractStringArrayArgument(named: "additionalMocksToGenerate", in: source)
-	}
-
-	/// Extracts `additionalDirectoriesToInclude` paths from a source file
-	/// containing `#SafeDIConfiguration`. Uses text-based scanning (no SwiftSyntax).
-	/// Returns an empty array if the file does not contain a configuration.
-	public static func extractAdditionalDirectoriesToInclude(in source: String) -> [String] {
-		extractStringArrayArgument(named: "additionalDirectoriesToInclude", in: source)
-	}
-
-	/// Extracts a string array argument value from a `#SafeDIConfiguration(...)` invocation.
-	/// Uses text-based scanning (no SwiftSyntax).
-	/// Returns an empty array if the argument is not found.
-	private static func extractStringArrayArgument(named argumentLabel: String, in source: String) -> [String] {
-		let sanitizedSource = sanitize(source: source)
-		let macroName = "#SafeDIConfiguration"
-		var macroSearchStart = sanitizedSource.startIndex
-		var configRange: Range<String.Index>?
-		while let candidateRange = sanitizedSource[macroSearchStart...].range(of: macroName) {
-			let afterMacro = candidateRange.upperBound
-			if afterMacro >= sanitizedSource.endIndex || !isIdentifierContinuation(sanitizedSource[afterMacro]) {
-				configRange = candidateRange
-				break
-			} else {
-				macroSearchStart = afterMacro
-			}
-		}
-		guard let configRange else { return [] }
-
-		// Find the opening parenthesis of the macro call.
-		var index = configRange.upperBound
-		skipWhitespace(in: sanitizedSource, index: &index)
-		guard index < sanitizedSource.endIndex,
-		      sanitizedSource[index] == "(",
-		      let closingParenIndex = matchingParenIndex(in: sanitizedSource, openingParenIndex: index)
-		else { return [] }
-
-		// Search for the argument label in the argument list.
-		let argumentList = sanitizedSource[sanitizedSource.index(after: index)..<closingParenIndex]
-		guard let labelRange = findArgumentLabel(named: argumentLabel, in: argumentList) else { return [] }
-
-		// Convert from sanitized-source index to original source to extract string content.
-		let colonOffset = sanitizedSource.distance(from: sanitizedSource.startIndex, to: labelRange.upperBound)
-		guard colonOffset < source.count else { return [] }
-		var originalIndex = source.index(source.startIndex, offsetBy: colonOffset)
-		skipWhitespace(in: source, index: &originalIndex)
-		guard originalIndex < source.endIndex, source[originalIndex] == ":" else { return [] }
-		originalIndex = source.index(after: originalIndex) // skip ':'
-
-		// Find the opening bracket of the array literal.
-		while originalIndex < source.endIndex, source[originalIndex] != "[" {
-			originalIndex = source.index(after: originalIndex)
-		}
-		guard originalIndex < source.endIndex else { return [] }
-
-		// Find the matching closing bracket.
-		var depth = 0
-		var closingIndex = originalIndex
-		while closingIndex < source.endIndex {
-			switch source[closingIndex] {
-			case "[": depth += 1
-			case "]":
-				depth -= 1
-				if depth == 0 {
-					let arrayContent = source[source.index(after: originalIndex)..<closingIndex]
-					return extractStringLiterals(from: arrayContent)
-				}
-			default: break
-			}
-			closingIndex = source.index(after: closingIndex)
-		}
-		return []
-	}
-
-	/// Finds a labeled argument at the top level of a macro argument list.
-	/// Returns the range of the label name if found, or `nil` if not present.
-	private static func findArgumentLabel(named label: String, in argumentList: Substring) -> Range<String.Index>? {
-		var parenthesisDepth = 0
-		var bracketDepth = 0
-		var braceDepth = 0
-		var searchStart = argumentList.startIndex
-
-		while searchStart < argumentList.endIndex {
-			switch argumentList[searchStart] {
-			case "(": parenthesisDepth += 1
-			case ")": parenthesisDepth -= 1
-			case "[": bracketDepth += 1
-			case "]": bracketDepth -= 1
-			case "{": braceDepth += 1
-			case "}": braceDepth -= 1
-			case _ where parenthesisDepth == 0 && bracketDepth == 0 && braceDepth == 0:
-				if argumentList[searchStart...].hasPrefix(label) {
-					let end = argumentList.index(searchStart, offsetBy: label.count)
-					if end >= argumentList.endIndex || !isIdentifierContinuation(argumentList[end]) {
-						return searchStart..<end
-					}
-				}
-			default:
-				break
-			}
-			searchStart = argumentList.index(after: searchStart)
-		}
-		return nil
-	}
-
-	/// Detects `@Instantiable(... generateMock: true ...)` in source text.
-	/// Returns `true` if any `@Instantiable` in the source has a `generateMock: true` argument.
-	public static func containsGenerateMockTrue(in source: String) -> Bool {
-		let sanitizedSource = sanitize(source: source)
-		let macroName = "@Instantiable"
-		var searchStart = sanitizedSource.startIndex
-
-		while let macroRange = sanitizedSource[searchStart...].range(of: macroName) {
-			var index = macroRange.upperBound
-			if index < sanitizedSource.endIndex,
-			   isIdentifierContinuation(sanitizedSource[index])
-			{
-				searchStart = index
-				continue
-			}
-
-			skipWhitespace(in: sanitizedSource, index: &index)
-			guard index < sanitizedSource.endIndex,
-			      sanitizedSource[index] == "(",
-			      let closingParenIndex = matchingParenIndex(
-			      	in: sanitizedSource,
-			      	openingParenIndex: index,
-			      )
-			else {
-				searchStart = macroRange.upperBound
-				continue
-			}
-
-			let arguments = sanitizedSource[sanitizedSource.index(after: index)..<closingParenIndex]
-			if containsGenerateMockTrueArgument(in: arguments) {
-				return true
-			}
-
-			searchStart = sanitizedSource.index(after: closingParenIndex)
-		}
-
-		return false
-	}
-
-	private static func containsGenerateMockTrueArgument(in arguments: Substring) -> Bool {
-		var clauseStart = arguments.startIndex
-		var parenthesisDepth = 0
-		var bracketDepth = 0
-		var braceDepth = 0
-		var index = arguments.startIndex
-
-		while index < arguments.endIndex {
-			switch arguments[index] {
-			case "(":
-				parenthesisDepth += 1
-			case ")":
-				parenthesisDepth -= 1
-			case "[":
-				bracketDepth += 1
-			case "]":
-				bracketDepth -= 1
-			case "{":
-				braceDepth += 1
-			case "}":
-				braceDepth -= 1
-			case "," where parenthesisDepth == 0 && bracketDepth == 0 && braceDepth == 0:
-				if isGenerateMockTrueClause(arguments[clauseStart..<index]) {
-					return true
-				}
-				clauseStart = arguments.index(after: index)
-			default:
-				break
-			}
-			index = arguments.index(after: index)
-		}
-
-		return isGenerateMockTrueClause(arguments[clauseStart..<arguments.endIndex])
-	}
-
-	private static func isGenerateMockTrueClause(_ clause: Substring) -> Bool {
-		let trimmedClause = clause.trimmingCharacters(in: .whitespacesAndNewlines)
-		guard trimmedClause.hasPrefix("generateMock") else { return false }
-
-		var index = trimmedClause.index(trimmedClause.startIndex, offsetBy: "generateMock".count)
-		if index < trimmedClause.endIndex,
-		   isIdentifierContinuation(trimmedClause[index])
-		{
-			return false
-		}
-
-		skipWhitespace(in: trimmedClause, index: &index)
-		guard index < trimmedClause.endIndex,
-		      trimmedClause[index] == ":"
-		else {
-			return false
-		}
-
-		index = trimmedClause.index(after: index)
-		skipWhitespace(in: trimmedClause, index: &index)
-
-		guard trimmedClause[index...].hasPrefix("true") else { return false }
-		index = trimmedClause.index(index, offsetBy: "true".count)
-		if index < trimmedClause.endIndex,
-		   isIdentifierContinuation(trimmedClause[index])
-		{
-			return false
-		}
-
-		skipWhitespace(in: trimmedClause, index: &index)
-		return index == trimmedClause.endIndex
-	}
-
-	/// Extracts quoted string literal values from the interior of an array expression.
-	private static func extractStringLiterals(from arrayContent: some StringProtocol) -> [String] {
-		var results = [String]()
-		var searchIndex = arrayContent.startIndex
-		while searchIndex < arrayContent.endIndex {
-			guard let openQuote = arrayContent[searchIndex...].firstIndex(of: "\"") else { break }
-			let contentStart = arrayContent.index(after: openQuote)
-			guard contentStart < arrayContent.endIndex,
-			      let closeQuote = arrayContent[contentStart...].firstIndex(of: "\"")
-			else { break }
-			results.append(String(arrayContent[contentStart..<closeQuote]))
-			searchIndex = arrayContent.index(after: closeQuote)
-		}
-		return results
-	}
-
-	private static func outputFileNames(
-		for inputURLs: [URL],
-		relativeTo baseURL: URL,
-		suffix: String = "+SafeDI.swift",
-	) -> [String] {
-		struct FileInfo {
-			let relativePath: String
-			let baseName: String
-			let parentComponents: [String]
-		}
-
-		let fileInfo = inputURLs.map { inputURL in
-			let relPath = relativePath(for: inputURL, relativeTo: baseURL)
-			let relativeDirectory = (relPath as NSString).deletingLastPathComponent
-			let parentComponents: [String] = if relativeDirectory.isEmpty || relativeDirectory == "." {
-				[]
-			} else {
-				relativeDirectory
-					.split(separator: "/")
-					.map(String.init)
-			}
-			return FileInfo(
-				relativePath: relPath,
-				baseName: inputURL.deletingPathExtension().lastPathComponent,
-				parentComponents: parentComponents,
-			)
-		}
-
-		var outputFileNames = Array(repeating: "", count: fileInfo.count)
-		let groups = Dictionary(grouping: Array(fileInfo.enumerated()), by: \.element.baseName)
-
-		for (baseName, entries) in groups {
-			guard entries.count > 1 else {
-				let entry = entries[0]
-				outputFileNames[entry.offset] = "\(baseName)\(suffix)"
-				continue
-			}
-
-			var namesByIndex = entries.reduce(into: [Int: String]()) { partialResult, entry in
-				partialResult[entry.offset] = baseName
-			}
-
-			var maxParentDepth = 0
-			for entry in entries {
-				maxParentDepth = max(maxParentDepth, entry.element.parentComponents.count)
-			}
-			if maxParentDepth > 0 {
-				for parentDepth in 1...maxParentDepth where Set(namesByIndex.values).count < entries.count {
-					for entry in entries {
-						let prefix = entry.element.parentComponents
-							.suffix(parentDepth)
-							.joined(separator: "_")
-						namesByIndex[entry.offset] = prefix.isEmpty ? baseName : "\(prefix)_\(baseName)"
-					}
-				}
-			}
-
-			for entry in entries {
-				let name = namesByIndex[entry.offset, default: baseName]
-				outputFileNames[entry.offset] = "\(name)\(suffix)"
-			}
-		}
-
-		return outputFileNames
-	}
-
-	private static func containsRootArgument(in arguments: Substring) -> Bool {
-		var clauseStart = arguments.startIndex
-		var parenthesisDepth = 0
-		var bracketDepth = 0
-		var braceDepth = 0
-		var index = arguments.startIndex
-
-		while index < arguments.endIndex {
-			switch arguments[index] {
-			case "(":
-				parenthesisDepth += 1
-			case ")":
-				parenthesisDepth -= 1
-			case "[":
-				bracketDepth += 1
-			case "]":
-				bracketDepth -= 1
-			case "{":
-				braceDepth += 1
-			case "}":
-				braceDepth -= 1
-			case "," where parenthesisDepth == 0 && bracketDepth == 0 && braceDepth == 0:
-				if isRootClause(arguments[clauseStart..<index]) {
-					return true
-				}
-				clauseStart = arguments.index(after: index)
-			default:
-				break
-			}
-			index = arguments.index(after: index)
-		}
-
-		return isRootClause(arguments[clauseStart..<arguments.endIndex])
-	}
-
-	private static func isRootClause(_ clause: Substring) -> Bool {
-		let trimmedClause = clause.trimmingCharacters(in: .whitespacesAndNewlines)
-		guard trimmedClause.hasPrefix("isRoot") else { return false }
-
-		var index = trimmedClause.index(trimmedClause.startIndex, offsetBy: "isRoot".count)
-		if index < trimmedClause.endIndex,
-		   isIdentifierContinuation(trimmedClause[index])
-		{
-			return false
-		}
-
-		skipWhitespace(in: trimmedClause, index: &index)
-		guard index < trimmedClause.endIndex,
-		      trimmedClause[index] == ":"
-		else {
-			return false
-		}
-
-		index = trimmedClause.index(after: index)
-		skipWhitespace(in: trimmedClause, index: &index)
-
-		guard trimmedClause[index...].hasPrefix("true") else { return false }
-		index = trimmedClause.index(index, offsetBy: "true".count)
-		if index < trimmedClause.endIndex,
-		   isIdentifierContinuation(trimmedClause[index])
-		{
-			return false
-		}
-
-		skipWhitespace(in: trimmedClause, index: &index)
-		return index == trimmedClause.endIndex
-	}
-
-	private static func matchingParenIndex(
-		in source: String,
-		openingParenIndex: String.Index,
-	) -> String.Index? {
-		var depth = 0
-		var index = openingParenIndex
-
-		while index < source.endIndex {
-			switch source[index] {
-			case "(":
-				depth += 1
-			case ")":
-				depth -= 1
-				if depth == 0 {
-					return index
-				}
-			default:
-				break
-			}
-			index = source.index(after: index)
-		}
-
-		return nil
-	}
-
-	private static func sanitize(source: String) -> String {
-		enum State {
-			case code
-			case lineComment
-			case blockComment(depth: Int)
-			case string(hashCount: Int, multiline: Bool)
-		}
-
-		var sanitized = ""
-		sanitized.reserveCapacity(source.count)
-		var state = State.code
-		var index = source.startIndex
-
-		while index < source.endIndex {
-			switch state {
-			case .code:
-				if hasPrefix("//", in: source, at: index) {
-					sanitized += "  "
-					index = source.index(index, offsetBy: 2)
-					state = .lineComment
-				} else if hasPrefix("/*", in: source, at: index) {
-					sanitized += "  "
-					index = source.index(index, offsetBy: 2)
-					state = .blockComment(depth: 1)
-				} else if let delimiter = stringDelimiterStart(in: source, at: index) {
-					sanitized += String(repeating: " ", count: delimiter.length)
-					index = source.index(index, offsetBy: delimiter.length)
-					state = .string(hashCount: delimiter.hashCount, multiline: delimiter.multiline)
-				} else {
-					sanitized.append(source[index])
-					index = source.index(after: index)
-				}
-
-			case .lineComment:
-				if source[index] == "\n" {
-					sanitized.append("\n")
-					index = source.index(after: index)
-					state = .code
-				} else {
-					sanitized.append(" ")
-					index = source.index(after: index)
-				}
-
-			case let .blockComment(depth):
-				if hasPrefix("/*", in: source, at: index) {
-					sanitized += "  "
-					index = source.index(index, offsetBy: 2)
-					state = .blockComment(depth: depth + 1)
-				} else if hasPrefix("*/", in: source, at: index) {
-					sanitized += "  "
-					index = source.index(index, offsetBy: 2)
-					state = depth == 1 ? .code : .blockComment(depth: depth - 1)
-				} else {
-					sanitized.append(source[index] == "\n" ? "\n" : " ")
-					index = source.index(after: index)
-				}
-
-			case let .string(hashCount, multiline):
-				if let delimiterLength = stringDelimiterEnd(
-					in: source,
-					at: index,
-					hashCount: hashCount,
-					multiline: multiline,
-				) {
-					sanitized += String(repeating: " ", count: delimiterLength)
-					index = source.index(index, offsetBy: delimiterLength)
-					state = .code
-				} else {
-					sanitized.append(source[index] == "\n" ? "\n" : " ")
-					index = source.index(after: index)
-				}
-			}
-		}
-
-		return sanitized
-	}
-
-	private struct StringDelimiter {
-		let hashCount: Int
-		let multiline: Bool
-		let length: Int
-	}
-
-	private static func stringDelimiterStart(
-		in source: String,
-		at index: String.Index,
-	) -> StringDelimiter? {
-		var hashCount = 0
-		var currentIndex = index
-
-		while currentIndex < source.endIndex, source[currentIndex] == "#" {
-			hashCount += 1
-			currentIndex = source.index(after: currentIndex)
-		}
-
-		guard currentIndex < source.endIndex,
-		      source[currentIndex] == "\""
-		else {
-			return nil
-		}
-
-		let multiline = hasPrefix("\"\"\"", in: source, at: currentIndex)
-		return StringDelimiter(
-			hashCount: hashCount,
-			multiline: multiline,
-			length: hashCount + (multiline ? 3 : 1),
-		)
-	}
-
-	private static func stringDelimiterEnd(
-		in source: String,
-		at index: String.Index,
-		hashCount: Int,
-		multiline: Bool,
-	) -> Int? {
-		if multiline {
-			guard hasPrefix("\"\"\"", in: source, at: index) else { return nil }
-			if hashCount == 0 {
-				guard !isEscapedQuote(in: source, at: index) else { return nil }
-				return 3
-			}
-			let hashStart = source.index(index, offsetBy: 3)
-			guard hasPrefix(String(repeating: "#", count: hashCount), in: source, at: hashStart) else {
-				return nil
-			}
-			return hashCount + 3
-		}
-
-		guard source[index] == "\"" else { return nil }
-		if hashCount == 0 {
-			guard !isEscapedQuote(in: source, at: index) else { return nil }
-			return 1
-		}
-		let hashStart = source.index(after: index)
-		guard hasPrefix(String(repeating: "#", count: hashCount), in: source, at: hashStart) else {
-			return nil
-		}
-		return hashCount + 1
-	}
-
-	private static func isEscapedQuote(in source: String, at index: String.Index) -> Bool {
-		var backslashCount = 0
-		var currentIndex = index
-
-		while currentIndex > source.startIndex {
-			let previousIndex = source.index(before: currentIndex)
-			guard source[previousIndex] == "\\" else { break }
-			backslashCount += 1
-			currentIndex = previousIndex
-		}
-
-		return !backslashCount.isMultiple(of: 2)
-	}
-
-	private static func hasPrefix(
-		_ prefix: String,
-		in source: String,
-		at index: String.Index,
-	) -> Bool {
-		source[index...].hasPrefix(prefix)
-	}
-
-	private static func skipWhitespace<S: StringProtocol>(
-		in source: S,
-		index: inout S.Index,
-	) {
-		while index < source.endIndex,
-		      source[index].isWhitespace
-		{
-			index = source.index(after: index)
-		}
-	}
-
-	private static func isIdentifierContinuation(_ character: Character) -> Bool {
-		character == "_" || character.isLetter || character.isNumber
-	}
-}
